requirements.txt
openai
langchain
langchain-openai
langchain-community
streamlit
python-dotenv
PyPDF2
pandas
ipykernel

-e .